requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.1.0
numpy==1.26.2

# Streamlit for testing/development interface
streamlit>=1.30.0

# HuggingFace Transformers for AI-generated image detection (needs 4.37+ for SigLIP support)
transformers>=4.40.0

# PyTorch CPU-only version (smaller, faster download)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
torchvision==0.17.0+cpu

# Optional additional ML libraries
# opencv-python==4.8.1.78
# scikit-learn==1.3.2
# onnxruntime==1.16.3
